ChocolateChip version 1.0.6 uses ECMAScript5.
This version uses EC5's Object.defineProperties method to extend objects. 

All methods and properties are now members of the base $ object. This includes the $$ method for retrieving element collections. To avoid having to write $.$$(), an alias is assigned at the end of the file so that you can simply use $$().

The $.getStyle method is removed. It was already deprecated in previous versions. Use $.style in its place.