Updated to version 1.1.2
Fixed element.delegate method to use isSameNode instead of isEqulNode to fix problem bypassing child elements that were event targets on complex controls.

Refactored $.extend to use Object.keys for browsers supporting ECMAScript5, with fall back for browsers that don't.

Added element aliases for body and app to reduce the number of searches for these. This allows you to do $.app.delegate() or $.body.css() or other such uses without having to get the nodes.

Add new method $.UICheckForOverflow to allow for checking if an element's content is overflowing or not. Can be used to check if scrolling is required.

Reorganized all methods into new order to reduce the number of extend operations being done internally.