Updated to version 1.2.0
Added new methods: Element.find and Element.findAll. These work like $
and $$ except that they start their search for their selectors based on
the element to which they are attached.
Revised all node related methods to always return "this" to allow
chaining of node related methods.
Added Element.prepend and Element.append to add content as the first or
last child of a collection. Their use is the same as their jQuery
equivalents.
The Element.css method has been refactored to handle arguments
differently. Now you can simply pass a sting in quotes without the
curly braces. You can also pass in an object literal of property/value
pairs. This was to make the arguments more like the usage of this
method in jQuery.
The Element.anim method was refactored to accept all arguments as an
object literal. This removes the necessity of remembering the proper
order for parameters.
Changed the eval statement in Element.xhrjson to use JSON.parse.
Added checks for iOS 4 and 5, as well as Safari 5.0 and 5.1.
Updated $.template to escape regex patterns.
Removed unnecessary var assignments.
Updated $.form2JSON for efficiency.
Changed how ChocolateChip passes its $ and $$ objects out to the global
namespace. If there is no $ and $$ in the global namespace,
ChocolateChip passes these out ass usual. It also passes out
$chocolatechip and $$chocolatechip. Using these you can sandbox your
code to prevent conflicts with other libraries that use $ and $$.