Multiple updates to ChocolateChip.js
Bumped to version 1.3.1

Refactored $.extend so that if only one argument is passed, the property to extend, it will be added to the $ object. This means you can write code like this to extend $: $.extend({newStuff: function(arg) { // Do stuff with arg... }});

Added new method: Array.hasNot. This is the opposite of Array.has in that it returns the nodes whose children do not match the argument.

Updated Array.prependTo to automatically reverse the order of elements being prepended so that they are in the same order as the original array.

Added new method $.html. This is just an alias for $.make to better indicate what it does.

Added new method $.isArray. This returns a boolean.

Added new method $.isFunction This returns a boolean.

Added new method $.isObject This returns a boolean.

Added new method Element.hasNot. This is the opposite of Element.has in that it returns the node if none of its children match the argument, otherwise it returns boolean false.

Added new method Element.html. This is the same as doing Element.empty().insert(content). It completely replaces the contents of the element with the new content.

Added new property $.userAction to make support for desktop and mobile user interaction easier. When loaded in a desktop environment this gets set to "click", whereas in mobile it gets set to "touchstart". On mobile there is a 300 millisecond delay for clicks, so this gives you the ability to provide clicks for desktop and touchstart for mobile with one property. You could use it when binding events like this: $("a").bind($.userAction, function() { //do stuff... });