Bumped ChocolateChip.js to version 1.3.3.
Updated $() to return document root when no selector is supplied, or the window object when window is passed as an argument.

Added cache system to allow storing, retrieving and deleting data in relation to a DOM node. This uses the following methods and properties:

- $.uuidNum
- $.makeUuid
- $.uuid
- $.chch_cache
- $.chch_cache.data
- $.chch_cache.data.keys
- $.chch_cache.data.values
- $.chch_cache.data.set
- $.chch_cache.data.get
- $.chch_cache.data.keyExists
- $.chch_cache.data.hasData
- $.chch_cache.data.delete
- Element.cache
- Element.uncache

Update Element.remove() to check the node for any cache and uncache it before deleting.